requirements.txt
# Darkstore – Tor onion store backend
# Python 3.11+ (Python 3.14: use fastapi<0.115 to avoid circular import in fastapi.applications)
fastapi>=0.109.0,<0.115.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
passlib[bcrypt]>=1.7.4
pyotp>=2.9.0
itsdangerous>=2.1.0
pillow>=10.0.0
python-dotenv>=1.0.0
